These following methods were changed:
- add(int): If the list was initially full, then
increase the size by 50% so there will be room.

- remove(int): If the list has more than 25% empty
places, the decrease the size of the list. The list
cannot be reduced to less than 1 entry.

These following methods were added:

- append(int): Appends the parameter to the end of the
list. If the list was full, then increase the size by
50% so there will be room.

- first(): Returns the first element in the list. If
there are no elements in the list, then return -1.

- last(): Returns the last element in the list. If
there are no elements in the list, then return -1..

- size(): Returns the current number of possible locations in the list.